develop <-- Patch/adjust item methods to take username
---
- accessToken checks were added from develop branch
- item endpoints now take String username (instead of int userId)
---
- import lines cleaned in all item java files
- header comment added all item java files

+++ src/main/java/com/wishlist/cst438project2/service/impl/AdminServiceImpl.java
@@ -194,7 +194,7 @@ public class AdminServiceImpl implements AdminService {
     /**
      * returns the item found in database by given name
      */
-    private Item fetchItem(String name, int userId) {
+    private Item fetchItem(String name, long userId) {
         ItemDTO dbItemDTO = firebaseIntegration.getItem(name, userId);
 
         if(Objects.isNull(dbItemDTO)) {
